Add dynamic and controllable trigger support
This moves control of the existing triggers, GPIO and signal, to
the playlist and makes them configurable.  Triggers are assigned per
sub-playlist or in the DEFAULTS scope.  Multiple triggers are allowed
per playlist, but at least one is required (for the playlist itself or
in any of its parents) for the playlist to become active.

The syntax to define a trigger is
 trigger=<player_id> <type> <arguments>

The player id is a user defined string the identifies which player this
trigger belongs to.  Multiple triggers can be mapped to the same player,
this allows one player to be controlled by multiple triggers.

To configure a GPIO trigger
 trigger=default gpio path=/path/to/gpio/pin

And to configure a signal trigger
 trigger=default singal toggle=<signum> [next=<signum>]
where signum can be sigusr1, sigusr2

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>