Add a random trigger
This trigger will randomly play a sound with configurable probability.

 trigger=<player> random delay=<seconds> dice=<roll>/<eyes>

Delay specifies how often the "dice" should be rolled.
Dice have two components, eyes are the total amount of eyes on the dice,
and roll specifies what counts as a win.  Everything upto and below
for <roll> counts as a win.

Examples

Will roll the dice once a second with a 1/6 probability of playing
 trigger=randomplayer delay=1 dice=1/6

Will roll the dice every 5 minutes with a 5/10 probability of playing
 trigger=randomplayer delay=300 dice=5/10

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>