Isolate each gstreamer instance to a separate process
Move each (relative) costly gstreamer player into separate processes,
since Python threads share the same execution context.  This allows
the gstream thread to run freely without the need to share execution
with the rest of the control code.  Control and state feedback
of each instance is performed over a pipe.

This also removes the need of a global watchdog as each gstreamer
instance will be monitored and restarted as needed.

Signed-off-by: Fredrik Lindberg <fli@shapeshifter.se>